NLP/requirements.txt
google-cloud-speech
google-api-python-client